reduce delta between stable and trunk:  merge branches/centreon-2.1@r8890-9560 into trunk
[ commits 9577-9584 doubled my work :( ]



git-svn-id: http://svn.centreon.com/trunk/centreon@9590 6bcd3966-0018-0410-8128-fd23d134de7e

+++ www/include/options/media/directory/DB-Func.php
@@ -76,6 +76,7 @@
 			$dir_alias =& $DBRESULT->fetchRow();
 			$fileTab = scandir("./img/media/" . $dir_alias["dir_alias"]);
 			foreach ($fileTab as $fileName)
+			    if ($filename != "." && $filename != "..")
 				unlink("./img/media/" . $dir_alias["dir_alias"] . "/" . $fileName);
 			rmdir("./img/media/".$dir_alias["dir_alias"]);
 			if (!is_dir("./img/media/".$dir_alias["dir_alias"]))	{
@@ -181,6 +182,7 @@
 		
 		$fileTab = scandir("./img/media/".$dir_alias);
 		foreach ($fileTab as $fileName) {
+		    if ($fileName != "." && $fileName != "..")
 			unlink("./img/media/".$dir_alias."/".$fileName);			
 		}
